requirements-dev.txt
-r requirements.txt

pytest==8.3.4
pytest-flask==1.3.0
mongomock==4.3.0
